requirements.txt
google-genai>=1.0.0
pymupdf>=1.24.0
edge-tts>=6.1.0
elevenlabs>=1.50.0
